build expanded view of overview image gallery
