+++ .eslintrc.json
@@ -14,6 +14,9 @@
     "react/jsx-uses-react": 1,
     "react/jsx-filename-extension": [1, { "extensions": [".js", ".jsx"] }],
     "react/prop-types": "off",
-    "react/self-closing-comp": "off"
+    "react/self-closing-comp": "off",
+    "react/jsx-props-no-spreading": ["off", {
+      "html": "ignore"
+  }]
   }
 }
