+++ client/src/components/common/Modal.jsx
@@ -1,0 +1,7 @@
+import ReactDOM from 'react-dom';
+
+function Modal({ children }) {
+  return ReactDOM.createPortal(children, document.querySelector('#modal-container'));
+}
+
+export default Modal;
