+++ client/src/components/overview/image-gallery/ExpandedView.jsx
@@ -1,0 +1,51 @@
+import React, { useState } from 'react';
+import styled from 'styled-components';
+import OverviewCarousel from './OverviewCarousel.jsx';
+import Modal from '../../common/Modal.jsx';
+import CloseIcon from '../../../assets/close-icon.png';
+
+const Host = styled.div`
+  width: 100vw;
+  height: 100vh;
+  background-color: rgba(0,0,0,0.6);
+`;
+
+const CloseButton = styled.div`
+  position: absolute;
+  top: 0;
+  right: 20px;
+  border-radius: 2px;
+  background-color: rgba(255,255,255,0.5);
+  &:hover {
+    box-shadow: 0 4px 8px 0 rgba(255,255,255,0.5);
+  }
+`;
+
+function ExpandedView({ photos, handleHide }) {
+  const [defaultViewIndex, setDefaultViewIndex] = useState(0);
+  const [thumbnailViewIndexStart, setThumbnailIndexStart] = useState(0);
+  const carouselProps = {
+    photos,
+    defaultViewIndex,
+    setDefaultViewIndex,
+    thumbnailViewIndexStart,
+    setThumbnailIndexStart,
+    inExpandedView: true,
+  };
+
+  return (
+    <Modal>
+      <Host>
+        <CloseButton onClick={() => {
+          handleHide(false);
+        }}
+        >
+          <img src={CloseIcon} alt="Close Logo" className="close-icon" />
+        </CloseButton>
+        <OverviewCarousel {...carouselProps} />
+      </Host>
+    </Modal>
+  );
+}
+
+export default ExpandedView;
